Port to 1.21.11

+++ src/main/java/cf/avicia/avomod2/client/eventhandlers/screenevents/TradeMarketMiddleClickSearch.java
@@ -30,8 +30,8 @@ public class TradeMarketMiddleClickSearch {
         new Thread(() -> {
             Utils.sendClickPacket(
                     screenHandler,
-                    47,
-                    0,
+                    (short) 47,
+                    (byte) 0,
                     SlotActionType.PICKUP,
                     searchItem
             );
